+++ src/agent_maintainer_kit/__init__.py
@@ -2,7 +2,7 @@
 
 from .checks import run_repo_checks
 from .policy import Policy, discover_policy_path, load_policy
-from .reporting import build_json_report, build_markdown_report, build_release_checklist
+from .reporting import build_json_report, build_markdown_report, build_release_checklist, build_review_comment
 from .transcript import analyze_transcript
 
 __all__ = [
@@ -11,6 +11,7 @@ __all__ = [
     "build_json_report",
     "build_markdown_report",
     "build_release_checklist",
+    "build_review_comment",
     "discover_policy_path",
     "load_policy",
     "run_repo_checks",